feat(core): WIP, implement basic game server

+++ sorapointa-core/src/main/kotlin/org/sorapointa/command/Command.kt
@@ -31,7 +31,7 @@ abstract class Command(
         val name: String,
         val help: String,
         val alias: List<String> = emptyList(),
-        val permissionRequired: UShort = 0u
+        val permissionRequired: Int = 0
     )
 
     init {
